rename handId folder to camel case
