+++ src/App.tsx
@@ -2,7 +2,7 @@ import Layout from './layout/Layout'
 import { Routes, Route, BrowserRouter } from 'react-router-dom'
 import Index from './pages/Index'
 import HandsIndex from './pages/hands/Index'
-import HandIdIndex from './pages/hands/_hand_id/Index'
+import HandIdIndex from './pages/hands/_handId/Index'
 import NotFound from './pages/NotFound'
 import { createContext } from 'react'
 import { Player } from './types'
