+++ src/pages/hands/_handId/components/AfterGame.tsx
