+++ src/pages/hands/_handId/components/BeforeGame.tsx
