+++ src/pages/hands/_handId/components/inputs/PartnerInput.tsx
