+++ src/pages/hands/_handId/helpers/helpers.ts
@@ -1,0 +1,8 @@
+export type SumParam = {
+  [index: string]: number
+}
+
+export function objSum(obj: SumParam) {
+  const values = Object.values(obj)
+  return values.reduce((acc, val) => acc + val)
+}
